move line actions to first field, instead of body field only
